requirements.txt
selenium==3.141.0
HtmlTestRunner==0.8.0
python-dotenv==0.18.0
